+++ models/staging/stg_geography_player_bins_all.sql
@@ -12,17 +12,17 @@ SELECT
 , SUM(player_bin_100) AS all_player_bin_100
 , SUM(player_bin_100_plus) AS all_player_bin_100_plus
 
-, SUM(player_bin_010) / SUM(tournament_count) AS all_player_bin_010_perc
-, SUM(player_bin_020) / SUM(tournament_count) AS all_player_bin_020_perc
-, SUM(player_bin_030) / SUM(tournament_count) AS all_player_bin_030_perc
-, SUM(player_bin_040) / SUM(tournament_count) AS all_player_bin_040_perc
-, SUM(player_bin_050) / SUM(tournament_count) AS all_player_bin_050_perc
-, SUM(player_bin_060) / SUM(tournament_count) AS all_player_bin_060_perc
-, SUM(player_bin_070) / SUM(tournament_count) AS all_player_bin_070_perc
-, SUM(player_bin_080) / SUM(tournament_count) AS all_player_bin_080_perc
-, SUM(player_bin_090) / SUM(tournament_count) AS all_player_bin_090_perc
-, SUM(player_bin_100) / SUM(tournament_count) AS all_player_bin_100_perc
-, SUM(player_bin_100_plus) / SUM(tournament_count) AS all_player_bin_100_plus_perc
+, SUM(player_bin_010::decimal) / SUM(tournament_count::decimal) AS all_player_bin_010_perc
+, SUM(player_bin_020::decimal) / SUM(tournament_count::decimal) AS all_player_bin_020_perc
+, SUM(player_bin_030::decimal) / SUM(tournament_count::decimal) AS all_player_bin_030_perc
+, SUM(player_bin_040::decimal) / SUM(tournament_count::decimal) AS all_player_bin_040_perc
+, SUM(player_bin_050::decimal) / SUM(tournament_count::decimal) AS all_player_bin_050_perc
+, SUM(player_bin_060::decimal) / SUM(tournament_count::decimal) AS all_player_bin_060_perc
+, SUM(player_bin_070::decimal) / SUM(tournament_count::decimal) AS all_player_bin_070_perc
+, SUM(player_bin_080::decimal) / SUM(tournament_count::decimal) AS all_player_bin_080_perc
+, SUM(player_bin_090::decimal) / SUM(tournament_count::decimal) AS all_player_bin_090_perc
+, SUM(player_bin_100::decimal) / SUM(tournament_count::decimal) AS all_player_bin_100_perc
+, SUM(player_bin_100_plus::decimal) / SUM(tournament_count::decimal) AS all_player_bin_100_plus_perc
 
 
 FROM {{ ref('stg_geography_player_bins') }} stg_geography_player_bins
